Fjernet Form og puttet ind i Login og ændret logo på forside og navbar
