--- planit-react/src/components/Form/Form.module.css
@@ -1,25 +1,0 @@
-.error_message {
-  color: red;
-  margin-top: 10px;
-}
-.question_button {
-  margin-top: 20px;
-  padding: 10px;
-  width: 10%;
-  border: 1px solid #050607;
-  border-radius: 5px;
-  background-color: #fafafa;
-  color: rgb(0, 0, 0);
-  font-weight: bold;
-}
-.question_button:hover {
-  background-color: #ebebeb;
-  color: #9ab693;
-  color: black;
-}
-.is_signed_up {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
